pyproject.toml
[tool.poetry]
name = "gpt-avsum"
version = "0.1.0"
description = ""
authors = ["Yauheni Bokach <dharmagetic@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.34.0"
python-dotenv = "^1.0.1"
pydub = "^0.25.1"
moviepy = "^1.0.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
